backend/core/requirements.txt
asgiref==3.8.1
cffi==2.0.0
colorama==0.4.6
coverage==7.11.0
cryptography==46.0.1
dj-database-url==3.0.1
Django==5.2.6
django-cors-headers==4.7.0
django-filter==25.1
django-restframework==0.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.8.5
drf-spectacular-sidecar==2025.10.1
gunicorn==23.0.0
iniconfig==2.3.0
packaging==25.0

pluggy==1.6.0
pycparser==2.23
Pygments==2.19.2
PyJWT==2.9.0
pytest==8.4.2
pytest-asyncio==1.2.0
python-dotenv==1.1.1
PyYAML==6.0.3
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
uv==0.8.22
whitenoise==6.9.0
 Pillow==10.4.0
psycopg[binary]>=3.2.2,<3.3
python-decouple==3.8
celery==5.3.4
redis==5.0.1
django-celery-beat==2.8.1
boto3==1.29.7
django-storages==1.14.2



